+++ tests/test_deprecated.py
@@ -14,6 +14,9 @@ def test_tbd_remove_in_v1_2_0():
     with pytest.deprecated_call(match='will be removed in v1.2'):
         checkpoint_cb = ModelCheckpoint(filepath='.')
 
+    with pytest.deprecated_call(match='will be removed in v1.2'):
+        checkpoint_cb = ModelCheckpoint('.')
+
     with pytest.raises(MisconfigurationException, match='inputs which are not feasible'):
         checkpoint_cb = ModelCheckpoint(filepath='.', dirpath='.')